+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### v0.3.0
+
+- Improve performance loading fixtures & expected points
+- Add total expected points to /team/my and /transfers
+- Add fixtures to /transfers sidebar and team
+- Add icons for transfers in/out to players, improve styling
+
 ### v0.2.0
 
 - Update classic league table